SpreadsheetEngineHateosHandlerSpreadsheetDeltaDelete was SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsert
