+++ src/test/java/walkingkooka/spreadsheet/server/engine/http/SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteColumnsTest.java
@@ -21,7 +21,6 @@ import org.junit.jupiter.api.Test;
 import walkingkooka.collect.Range;
 import walkingkooka.collect.list.Lists;
 import walkingkooka.collect.map.Maps;
-import walkingkooka.collect.set.Sets;
 import walkingkooka.net.http.server.hateos.HateosHandler;
 import walkingkooka.spreadsheet.SpreadsheetCell;
 import walkingkooka.spreadsheet.engine.FakeSpreadsheetEngine;
@@ -38,25 +37,25 @@ import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertDeleteRowsTest extends SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertTestCase<SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertDeleteRows,
-        SpreadsheetRowReference> {
+public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteColumnsTest extends SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteTestCase<SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteColumns,
+        SpreadsheetColumnReference> {
 
     @Test
-    public void testDeleteRow() {
-        final SpreadsheetRowReference row = this.id();
+    public void testDeleteColumn() {
+        final SpreadsheetColumnReference column = this.id();
         final Optional<SpreadsheetDelta> resource = this.resource();
 
-        final Set<SpreadsheetCell> cells = Sets.of(this.cell());
+        final Set<SpreadsheetCell> cells = this.cells();
 
         this.handleOneAndCheck(this.createHandler(
                 new FakeSpreadsheetEngine() {
 
                     @Override
                     @SuppressWarnings("OptionalGetWithoutIsPresent")
-                    public SpreadsheetDelta deleteRows(final SpreadsheetRowReference r,
-                                                       final int count,
-                                                       final SpreadsheetEngineContext context) {
-                        assertEquals(row, r, "row");
+                    public SpreadsheetDelta deleteColumns(final SpreadsheetColumnReference c,
+                                                          final int count,
+                                                          final SpreadsheetEngineContext context) {
+                        assertEquals(column, c, "column");
                         assertEquals(1, count, "count");
                         return SpreadsheetDelta.EMPTY.setCells(cells);
                     }
@@ -70,11 +69,10 @@ public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertD
                     @Override
                     public double rowHeight(final SpreadsheetRowReference row,
                                             final SpreadsheetEngineContext context) {
-                        assertEquals(SpreadsheetRowReference.parseRow("99"), row, "row");
                         return ROW_HEIGHT.pixelValue();
                     }
                 }),
-                row,
+                column,
                 resource,
                 HateosHandler.NO_PARAMETERS,
                 Optional.of(
@@ -82,7 +80,8 @@ public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertD
                                 .setCells(cells)
                                 .setColumnWidths(
                                         Maps.of(
-                                                SpreadsheetColumnReference.parseColumn("A"), COLUMN_WIDTH.pixelValue()
+                                                SpreadsheetColumnReference.parseColumn("A"), COLUMN_WIDTH.pixelValue(),
+                                                SpreadsheetColumnReference.parseColumn("Z"), COLUMN_WIDTH.pixelValue()
                                         )
                                 )
                                 .setRowHeights(
@@ -95,10 +94,10 @@ public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertD
     }
 
     @Test
-    public void testDeleteSeveralRows() {
+    public void testDeleteSeveralColumns() {
         final Optional<SpreadsheetDelta> resource = this.collectionResource();
 
-        final Range<SpreadsheetRowReference> range = SpreadsheetColumnOrRowReference.parseRowRange("2:4")
+        final Range<SpreadsheetColumnReference> range = SpreadsheetColumnOrRowReference.parseColumnRange("C:E")
                 .range();
         final Set<SpreadsheetCell> cells = this.cells();
 
@@ -108,36 +107,43 @@ public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertD
                 new FakeSpreadsheetEngine() {
 
                     @Override
-                    public SpreadsheetDelta deleteRows(final SpreadsheetRowReference r,
-                                                       final int count,
-                                                       final SpreadsheetEngineContext context) {
-                        assertEquals(SpreadsheetColumnOrRowReference.parseRow("2"), r, "row");
-                        assertEquals(3, count, "count"); // 2, 3 & 4
+                    public SpreadsheetDelta deleteColumns(final SpreadsheetColumnReference c,
+                                                          final int count,
+                                                          final SpreadsheetEngineContext context) {
+                        assertEquals(SpreadsheetColumnOrRowReference.parseColumn("C"), c, "column");
+                        assertEquals(3, count, "count"); // C, D & E
                         return delta;
                     }
 
                     @Override
                     public double columnWidth(final SpreadsheetColumnReference column,
                                               final SpreadsheetEngineContext context) {
+                        switch (column.toString()) {
+                            case "A":
+                            case "Z":
+                                break;
+                            default:
+                                throw new UnsupportedOperationException("Unexpected column: " + column);
+                        }
+
                         return 0;
                     }
 
                     @Override
                     public double rowHeight(final SpreadsheetRowReference row,
                                             final SpreadsheetEngineContext context) {
-                        assertEquals(SpreadsheetRowReference.parseRow("99"), row, "row");
                         return 0;
                     }
                 }),
-                range, // 3 inclusive
+                range, // 2 inclusive
                 resource,
                 HateosHandler.NO_PARAMETERS,
                 Optional.of(delta));
     }
 
     @Test
-    public void testDeleteRowFiltered() {
-        final SpreadsheetRowReference row = this.id();
+    public void testDeleteColumnFiltered() {
+        final SpreadsheetColumnReference column = this.id();
 
         final Set<SpreadsheetCell> cells = this.cells();
         final Optional<SpreadsheetCellRange> window = this.window();
@@ -148,10 +154,10 @@ public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertD
 
                             @Override
                             @SuppressWarnings("OptionalGetWithoutIsPresent")
-                            public SpreadsheetDelta deleteRows(final SpreadsheetRowReference c,
-                                                               final int count,
-                                                               final SpreadsheetEngineContext context) {
-                                assertEquals(row, c, "row");
+                            public SpreadsheetDelta deleteColumns(final SpreadsheetColumnReference c,
+                                                                  final int count,
+                                                                  final SpreadsheetEngineContext context) {
+                                assertEquals(column, c, "column");
                                 assertEquals(1, count, "count");
                                 return SpreadsheetDelta.EMPTY.setCells(cells);
                             }
@@ -159,17 +165,18 @@ public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertD
                             @Override
                             public double columnWidth(final SpreadsheetColumnReference column,
                                                       final SpreadsheetEngineContext context) {
+                                assertEquals(SpreadsheetColumnReference.parseColumn("A"), column, "column");
                                 return 0;
                             }
 
                             @Override
-                            public double rowHeight(final SpreadsheetRowReference r,
+                            public double rowHeight(final SpreadsheetRowReference row,
                                                     final SpreadsheetEngineContext context) {
-                                assertEquals(SpreadsheetRowReference.parseRow("99"), r);
                                 return 0;
                             }
-                        }),
-                row,
+                        }
+                ),
+                column,
                 Optional.empty(),
                 Maps.of(
                         SpreadsheetEngineHateosHandlerSpreadsheetDelta.WINDOW, Lists.of(window.get().toString())
@@ -183,23 +190,23 @@ public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertD
     }
 
     @Test
-    public void testDeleteAllRowsFails() {
+    public void testDeleteAllColumnsFails() {
         this.handleRangeFails2(Range.all());
     }
 
     @Test
     public void testDeleteOpenRangeBeginFails() {
-        this.handleRangeFails2(Range.lessThanEquals(SpreadsheetColumnOrRowReference.parseRow("2")));
+        this.handleRangeFails2(Range.lessThanEquals(SpreadsheetColumnOrRowReference.parseColumn("A")));
     }
 
     @Test
     public void testDeleteOpenRangeEndFails() {
-        this.handleRangeFails2(Range.greaterThanEquals(SpreadsheetColumnOrRowReference.parseRow("3")));
+        this.handleRangeFails2(Range.greaterThanEquals(SpreadsheetColumnOrRowReference.parseColumn("A")));
     }
 
-    private void handleRangeFails2(final Range<SpreadsheetRowReference> rows) {
-        assertEquals("Range with both rows required=" + rows,
-                this.handleRangeFails(rows,
+    private void handleRangeFails2(final Range<SpreadsheetColumnReference> columns) {
+        assertEquals("Range with both columns required=" + columns,
+                this.handleRangeFails(columns,
                         this.collectionResource(),
                         HateosHandler.NO_PARAMETERS,
                         IllegalArgumentException.class).getMessage(),
@@ -208,32 +215,32 @@ public final class SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertD
 
     @Test
     public void testToString() {
-        this.toStringAndCheck(this.createHandler().toString(), "SpreadsheetEngine.deleteRows");
+        this.toStringAndCheck(this.createHandler().toString(), "SpreadsheetEngine.deleteColumns");
     }
 
-    private SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertDeleteRows createHandler(final SpreadsheetEngine engine) {
+    private SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteColumns createHandler(final SpreadsheetEngine engine) {
         return this.createHandler(engine, this.engineContext());
     }
 
     @Override
-    SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertDeleteRows createHandler(final SpreadsheetEngine engine,
-                                                                                         final SpreadsheetEngineContext context) {
-        return SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertDeleteRows.with(engine, context);
+    SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteColumns createHandler(final SpreadsheetEngine engine,
+                                                                              final SpreadsheetEngineContext context) {
+        return SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteColumns.with(engine, context);
     }
 
     @Override
-    public SpreadsheetRowReference id() {
-        return SpreadsheetColumnOrRowReference.parseRow("2");
+    public SpreadsheetColumnReference id() {
+        return SpreadsheetColumnOrRowReference.parseColumn("C");
     }
 
     @Override
-    public Range<SpreadsheetRowReference> range() {
-        return SpreadsheetColumnOrRowReference.parseRowRange("2:4")
+    public Range<SpreadsheetColumnReference> range() {
+        return SpreadsheetColumnOrRowReference.parseColumnRange("C:E")
                 .range();
     }
 
     @Override
-    public Class<SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertDeleteRows> type() {
-        return SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteOrInsertDeleteRows.class;
+    public Class<SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteColumns> type() {
+        return SpreadsheetEngineHateosHandlerSpreadsheetDeltaDeleteColumns.class;
     }
 }
